+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### [1.3.28] - Unreleased
 - Add Color::from_rgba and to_rgba with enable-glwindow feature.
 - Make Color::from_hex, from_hex_str, to_hex_str RGBA aware.
+- Add Tabs::auto_layout().
 
 ### [1.3.27] - 2023-01-14
 - Pull FLTK fixes.