+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
 
-### [1.3.28] - Unreleased
+### [1.3.28] - 2023-01-23
 - Add image::AnimGifImage.
+- Add Tabs::auto_layout().
 - Add Color::from_rgba and to_rgba with enable-glwindow feature.
 - Make Color::from_hex, from_hex_str, to_hex_str RGBA aware.
-- Add Tabs::auto_layout().
 - Expose internal utils::oncelock module.
 
 ### [1.3.27] - 2023-01-14